requirements.txt
pyodbc
flask
gunicorn
speechrecognition
ffmpeg-python
werkzeug